+++ NEWS.md
@@ -4,6 +4,8 @@ output: html_document
 ---
 
 
+* `footporint_aggregates()` now also includes the
+  `*_prime_colname` matrices in the nested data frame.
 * `calc_io_mats()` and `calc_L()`
   gain `method` and `tol` arguments
   to control matrix inversion.